+++ examples/nextjs-shopify/src/storefront/hooks/index.tsx
@@ -30,7 +30,7 @@ export const {
   useRemoveLineItemsFromCartMutation,
   useUpdateLineItemsInCartMutation
 } = createStorefrontHooks({
-  cartLocalStorageKey: 'example-nextjs-shopify',
+  cartCookieKey: 'example-nextjs-shopify',
   fetchers: {
     fetchCart: async (cartId) => {
       const { cart } = await storefront.query({